Replace SyncTimeManagerImpl with GenresSyncTimeManager and PlatformsSyncTimeManager. Refactor GenresRepositoryImpl, PlatformsRepositoryImpl, UserPreferencesRepositoryImpl
