+++ core/data/src/main/java/ru/d3rvich/data/repositories/GenresRepositoryImpl.kt
@@ -1,9 +1,8 @@
-package ru.d3rvich.data.repositoties
+package ru.d3rvich.data.repositories
 
 import kotlinx.coroutines.flow.Flow
+import org.koin.core.annotation.ComponentScan
 import org.koin.core.annotation.Factory
-import org.koin.core.annotation.Named
-import org.koin.core.annotation.Qualifier
 import ru.d3rvich.core.domain.entities.GenreFullEntity
 import ru.d3rvich.core.domain.model.LoadingResult
 import ru.d3rvich.core.domain.model.map
@@ -11,6 +10,7 @@ import ru.d3rvich.core.domain.repositories.GenresRepository
 import ru.d3rvich.data.mapper.asResult
 import ru.d3rvich.data.mapper.toGenreDBO
 import ru.d3rvich.data.mapper.toGenreFullEntity
+import ru.d3rvich.data.model.GenresSync
 import ru.d3rvich.data.model.SyncTimeManager
 import ru.d3rvich.data.model.localDataSource
 import ru.d3rvich.data.util.cashedRemoteRequest
@@ -21,11 +21,12 @@ import ru.d3rvich.remote.util.getAllGenres
 /**
  * Created by Ilya Deryabin at 04.04.2024
  */
-@Factory
+@Factory(binds = [GenresRepository::class])
+@ComponentScan("ru.d3rvich.data.model")
 internal class GenresRepositoryImpl(
     private val apiService: JetGamesNetworkDataSource,
     private val database: JetGamesDatabase,
-    @Named("Genres") private val syncTimeManager: SyncTimeManager,
+    @param:GenresSync private val syncTimeManager: SyncTimeManager,
 ) : GenresRepository {
 
     override fun getGenres(): Flow<LoadingResult<List<GenreFullEntity>>> {
@@ -36,7 +37,8 @@ internal class GenresRepositoryImpl(
             syncTimeManager = syncTimeManager,
             localDataSource = localDataSource,
             remoteCall = {
-                apiService.getAllGenres().asResult().map { list -> list.map { it.toGenreFullEntity() } }
+                apiService.getAllGenres().asResult()
+                    .map { list -> list.map { it.toGenreFullEntity() } }
             }
         )
     }
